修改类型

+++ eladmin-system/src/main/java/me/zhengjie/modules/Archivesmouthsmanagement/repository/ArchivesmouthsmanagementRepository.java
@@ -4,5 +4,5 @@ import me.zhengjie.modules.Archivesmouthsmanagement.domain.Archivesmouthsmanagem
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface ArchivesmouthsmanagementRepository extends JpaRepository<Archivesmouthsmanagement, Integer>, JpaSpecificationExecutor {
+public interface ArchivesmouthsmanagementRepository extends JpaRepository<Archivesmouthsmanagement, Long>, JpaSpecificationExecutor {
 }
